requirements.txt
# Core dependencies
requests>=2.31.0

# Web application dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0

# Optional: for better development experience
python-dotenv>=1.0.0
